+++ backend/src/baserow/contrib/database/api/views/errors.py
@@ -1,0 +1,8 @@
+from rest_framework.status import HTTP_404_NOT_FOUND
+
+
+ERROR_VIEW_DOES_NOT_EXIST = (
+    'ERROR_VIEW_DOES_NOT_EXIST',
+    HTTP_404_NOT_FOUND,
+    'The requested view does not exist.'
+)