+++ backend/tests/baserow/contrib/database/field/test_field_handler.py
@@ -7,8 +7,43 @@ from baserow.contrib.database.fields.models import (
     Field, TextField, NumberField, BooleanField
 )
 from baserow.contrib.database.fields.exceptions import (
-    FieldTypeDoesNotExist, PrimaryFieldAlreadyExists, CannotDeletePrimaryField
+    FieldTypeDoesNotExist, PrimaryFieldAlreadyExists, CannotDeletePrimaryField,
+    FieldDoesNotExist
 )
+
+
+@pytest.mark.django_db
+def test_get_field(data_fixture):
+    user = data_fixture.create_user()
+    user_2 = data_fixture.create_user()
+    text = data_fixture.create_text_field(user=user)
+
+    handler = FieldHandler()
+
+    with pytest.raises(FieldDoesNotExist):
+        handler.get_field(user=user, field_id=99999)
+
+    with pytest.raises(UserNotInGroupError):
+        handler.get_field(user=user_2, field_id=text.id)
+
+    field = handler.get_field(user=user, field_id=text.id)
+
+    assert text.id == field.id
+    assert text.name == field.name
+    assert isinstance(field, Field)
+
+    field = handler.get_field(user=user, field_id=text.id, field_model=TextField)
+
+    assert text.id == field.id
+    assert text.name == field.name
+    assert isinstance(field, TextField)
+
+    # If the error is raised we know for sure that the query has resolved.
+    with pytest.raises(AttributeError):
+        handler.get_field(
+            user=user, field_id=text.id,
+            base_queryset=Field.objects.prefetch_related('UNKNOWN')
+        )
 
 
 @pytest.mark.django_db